+++ CHANGELOG.md
@@ -4,6 +4,9 @@
     - update plugins.Analysis.Stats to allow user to dynamically change whether 
       to process externally loaded spectra
     - update plugins.Analysis.Peakfinding with more params and copy button
+    - Marquee 
+        - made threadsafe for plugins
+        - removed toast
 - 2024-02-11 4.1.22
     - updated Wasatch.PY to 2.2.17 to resolve broken Batch Collection and TakeOne buttons
     - plugins